requirements.txt
pycairo>=1.25.0
Pillow>=10.0.0
numpy>=1.26.0
